Use note field from propoisition model
I've added a note field in a previous commit. It sole purpose at the
moment is to have written somewhere why i've denied a suggestion. so i'm
using it to tell the user if a suggestion is denied, why it does not
appear. Also, bumping proposition limit because there's not enough
user to spam too much at the moment.